Update  modulo secciones dinamicas en grados y otros
**Se corrigio el error de mostrar secciones que no estaban activas al consultar secciones disponible por grado al asignar profesores en un grado

**Se añadio un estado de carga a los select que cargen infomacion por ejemplo selects municipios y parroquias

**Se corrigio la accion cuando se restaura una bd, ahora cerrara la sesion y tendra que a volver ingresar al sistema

**Se corrigio la vista de la tabla usuario para no poder eliminar un administrador

**Correciones de seguridad para no permitir la eliminacion de un administrado

+++ editar/eliminar_usuario.php
@@ -1,17 +1,19 @@
 <?php
 include("../db.php");
 
+if ($_SESSION['cargo'] != 1) {
+
     $id_usuario = $_GET['id'];
 
     $elimar_usuario = "DELETE FROM `usuario` WHERE usuario='$id_usuario'";
     $ejecutar_eliminar_usuario = mysqli_query($conexion, $elimar_usuario);
 
-    if($ejecutar_eliminar_usuario){ 
-    $resultado=["resultado"=>"exito"];
-    echo json_encode($resultado);
+    if ($ejecutar_eliminar_usuario) {
+        $resultado = ["resultado" => "exito"];
+        echo json_encode($resultado);
+    } else {
 
-   } else {
- 
-    $resultado=["resultado"=>"error"];
-    echo json_encode($resultado);
+        $resultado = ["resultado" => "error"];
+        echo json_encode($resultado);
     }
+}
